+++ src/lib/task-callback.ts
@@ -1,5 +1,5 @@
 import { SimpleGitTask } from './tasks/task';
-import { GitError } from './errors/git-error';
+import { GitError, GitResponseError } from './api';
 import { NOOP } from './utils';
 import { SimpleGitTaskCallback } from './types';
 
@@ -9,12 +9,41 @@ export function taskCallback<R>(task: SimpleGitTask<R>, response: Promise<R>, ca
       callback(null, data);
    };
 
-   const onError = (err: GitError) => {
+   const onError = (err: GitError | GitResponseError) => {
       if (err?.task === task) {
+         if (err instanceof GitResponseError) {
+            return callback(addDeprecationNoticeToError(err));
+         }
          callback(err);
       }
    };
 
    response.then(onSuccess, onError);
 
+}
+
+function addDeprecationNoticeToError (err: GitResponseError) {
+   let log = (name: string) => {
+      console.warn(`simple-git deprecation notice: accessing GitResponseError.${name} should be GitResponseError.git.${name}`);
+      log = NOOP;
+   };
+
+   return Object.create(err, Object.getOwnPropertyNames(err.git).reduce(descriptorReducer, {}));
+
+   function descriptorReducer(all: PropertyDescriptorMap, name: string): typeof all {
+      if (name in err) {
+         return all;
+      }
+
+      all[name] = {
+         enumerable: false,
+         configurable: false,
+         get () {
+            log(name);
+            return err.git[name];
+         },
+      };
+
+      return all;
+   }
 }