+++ CHANGELOG.md
@@ -4,15 +4,35 @@
 <!-- Notes added below this line -->
 <!-- Template: ${version} -->
 
-## 2.6.0 - Native Promises, TypeScript Importing and Git.clean
+## 2.6.0 - Native Promises, Typed Errors, TypeScript Importing and Git.clean
 
+### Native Promises
+
+- _TL;DR - `.then` and `.catch` can now be called on the standard `simpleGit` chain to handle the promise
+  returned by the most recently added task... essentially, promises now just work the way you would expect
+  them to._
 - The main export from `simple-git` no longer shows the deprecation notice for using the
   `.then` function, it now exposes the promise chain generated from the most recently run
   task, allowing the combination of chain building and ad-hoc splitting off to a new promise chain.
   - See the [unit](./test/unit/promises.spec.js) and [integration](./test/integration/promise-from-root.spec.js) tests.
   - See the [typescript consumer](./test/consumer/ts-default-from-root.spec.ts) test.
+
+### TypeScript Importing
+
 - Promise / async interface and TypeScript types all available from the `simple-git` import rather than needing
   `simple-git/promise`, see examples in the [ReadMe](./readme.md) or in the [consumer tests](./test/consumer).
+
+### Typed Errors
+
+- Tasks that previously validated their usage and rejected with a `TypeError` will now reject with a
+ [`TaskConfigurationError`](./src/lib/errors/task-configuration-error.ts).
+
+- Tasks that previously rejected with a custom object (currently only `git.merge` when the auto-merge fails)
+  will now reject with a [`GitResponseError`](./src/lib/errors/git-response-error.ts) where previously it
+  was a modified `Error`.
+
+### Git Clean
+
 - `git.clean(...)` will now return a `CleanSummary` instead of the raw string data
 
 ## 2.5.0 - Git.remote